Updating userForProfile dto to include email and phone

+++ FullStackAuth_WebAPI/DataTransferObjects/UserProfileWithReviewSummaryDto.cs
@@ -4,14 +4,14 @@ namespace FullStackAuth_WebAPI.DataTransferObjects
 {
     public class UserProfileWithReviewSummaryDto
     {
-        public int TotalReviewsJobs { get; set; }
-        public int AvgOverallScore { get; set; }
-        public int AvgAdherence { get; set;}
-        public int AvgQuality { get; set;}
-        public int AvgTimeliness { get; set; }
-        public int AvgWouldRepeat { get; set; }
-        public int AvgCommunication { get; set; }
-        public int? AvgAdaptability { get; set; }
+        public double TotalReviewsJobs { get; set; }
+        public double AvgOverallScore { get; set; }
+        public double AvgAdherence { get; set;}
+        public double AvgQuality { get; set;}
+        public double AvgTimeliness { get; set; }
+        public double AvgWouldRepeat { get; set; }
+        public double AvgCommunication { get; set; }
+        public double? AvgAdaptability { get; set; }
         public bool IsWorker { get; set; }
 
         public UserForProfileDto User { get; set; }
